.emacs.el: datetime-parse: support "short" parsing
Adds the option to parse lower-precision inputs to
just the short ISO8601 datetime which only contains
the parts which were specified by the input, without
the "floor fill" of lowest values.

So for example "Jun" can now be optionally parsed
to just "202406" instead of "20240601T000000", "May
5" can be parsed to just "20250505", "5 30 sun" to
"20240526T0530", and so on.

This enables interfaces which need to disambiguate
what precision the user actually wanted, such as
"give me a list of all appointments in June" vs
"give me a list of all appointments on June 1st".