.emacs.el: fixed-decoded-time-add: keep dst and zone
`encode-time` assumes that a decoded-time list will
have all of its fields, which finally forces me to
think about what the right action for DST and time
zone information is on datetime arithmetic.

So long as both are unchanged, timezone and DST
can be passed through unchanged and remain valid,
since both of these pieces of information are
just offsets: so long as they remain constant,
even if they're no longer the canonical way to
represent time in any given jurisdiction after
the add, they're still representing the correct
point in time.