.emacs.el: draft of switch-to-buffer-{next,previous}
Abandoning this because the behavior just isn't nice
enough to bother with, thanks to... Idek, probably
some deep Emacs workings of how buffers are ordered.

For example, with this implementation, with eight of
my `*eshell: git ...*` buffers matched with the last
search "git", this cycles though... five of them.
Not eight, five. I don't know how much Emacs' "bury"
and so on are involved. Too much for me to learn at
this time.