.emacs.el: window-state: buffer-switching "/"/"n"/"N"
Mnemonic: in vi, the "/" operator searches, so the
window state operator "/" searches for a thing to
display.

This is largely redundant with the "c" operator, but
I want to see if I like the two-keypress convenience.
(I currently have " b" to do the same thing though,
with muscle memory of full-time use for a year behind
it, so the competition is pretty strong for "s/". But
it would be nice to get all window management actions
into my window state, and then that's one or more
space-map key for other stuff).

In the long run, I might consider making "/" a
full operator the way "c" is now, so that it takes
"s/{{motion}}" or "s//" to start a search, instead
of just "s/". (This is part of why I'm leaving "c"
as-is for now - if I start wanting the full "/", I
have "c" still on-hand to get a feel for that UX.
This would especially make sense if I keep " b",
since then it's two keybinds where " b" is shortest
for the current window, and "s/{{motion}}" is
shortest for any other window.)

However, I think that extra keypress (or two - f.e.
"scL": the "L" costs a Shift press and an "l" press)
is the main reason why I don't use the "c" operator.
But now that window-managing operators select their
target window and the search is integrated with
`switch-to-buffer-{next/previous}`, maybe I'll see
more use for the full operator (since "s//" in the
"switch buffer in current window" case instead of
"s/" is a small price to pay if I end up using it
more often in combination with creating/switching
to other window splits).