.emacs.el: make " a" same as " A"
Longer-term, I probably want some integration of
the two, like a calendar view that shows that
day's task list buffer as I move through the days.
That's why I put calendar on "a" to begin with:
a placeholder for that kind of vision.

But for now, I'm not using the calendar, but I
_am_ using my "agenda" view (task list buffer for
today or tomorrow), and I haven't yet assigned
"a" to anything else (like the idea to eventually
have " a" create scheduled task, like I currently
have "  g" doing).