.emacs.el: make minibuffer shrink properly again
At some point, an update caused the minibuffer
to wrongly not shrink in some cases where it
used to and really should. For example, Eshell
`sudo` commands invoked from my `history-execute`
should shrink the minibuffer down to one line for
the password prompt - instead, it kept the much
larger minibuffer size that Vertico uses.

Turns out this is because the default value of
`resize-mini-windows` is `grow-only`, not `t`,
and Vertico also unconditionally sets it to
`grow-only` rather than `t`. I'm not sure if
the latter alone would be enough to solve this
problem, but changing it globally is (based on
initial testing of the Eshell sudo use-case),
and frankly I think `grow-only` is wrong/worse
UX in the general case.