.emacs.el: switch to narrowing task buffers again
Turns out there is a bug in Dired and/or Emacs which
locks up some operations (f.e., `move-to-window-line`)
when the top line in a Dired buffer has been deleted
(as my optimization did instead of narrowing), the
Dired buffer is otherwise empty, and a single space
is added to the buffer (as ace-window does to draw
the target number hints - ace-window could be hacked
or changed upstream to not do this, but it might be
better for ace-window to ignore this since really it
should be fixed upstream in Emacs instead, and even
if ace-window works around it, in principle this bug
can be triggered in other ways).

So the optimization is out. Luckily, it doesn't seem
to be necessary anymore, thanks to the optimized
fontification for Denoted names. This also lets us
clean out the other workaround for marking the
first line in Task buffers.