.emacs.el: consult-line-next: fix rapid calls
For reasons I haven't fully dug to the bottom of,
holding down the consult-line-next keybind, the
search would fail.

We can inject keys into `unread-command-events`
to reproduce, for example, with my vi-style
binding of "n" for the next search result:

    (setq unread-command-events
          (append (listify-key-sequence "nn")
                  unread-command-events))

Forcing a Vertico update (sadly, need to call a
private function, no stable API for it) before
checking if there's no matches seems to fix the
problem.