.emacs.el: command-execute-in-keymap: drop read-key-sequence arguments
These arguments to `read-key-sequence` never seem
to be what I want/need, and whenever I see the
signature of `command-execute-in-keymap`, I can't
even remember what they are for. The latter can
be solved by changing the argument name, but the
former strongly suggests it's time to drop it and
see if I ever end up needing it.