.emacs.el: command-execute-in-keymap: make running hooks optional
Running the hooks is wrong when the
`command-execute-in-keymap` invocation
is meant to be part of the command
that's calling it.

Not running the hooks is wrong when
the invocation is meant to behave
like a full command in its own right.

I'm implementing this as an optional
parameter because a separate function
would need either a lot of duplication
or would depend on extra complexity in
`run-{pre,post}-command-hook`.

Since `command-execute` doesn't run
hooks normally, I've decided to make
the default behavior match that.