.emacs.el: window-state: refactor to use keymap
The idiomatic way for binding commands to keys
in Emacs is keymaps. For some reason, ace-window
uses a raw alist instead. I presume it was just
easier and more mentally accessible or something.
But it means you can't reuse any of the existing
tooling and code for managing and using keybinds.

In this commit, I finally get around to making
the change to using a normal keymap instead of
the `aw-dispatch-alist`.