.emacs.el: Russian keybindings
Just the basics for now, the keys that don't need
any non-trivial collision-resolving decisions,
and just for keymaps where the mapping is trivial.

The tricky thing with mapping vi-style bindings to
the standard Russian keyboard layouts is that
Cyrillic has so many symbols that all the symbols
end up shifted around or entirely unavailable.

So for example, the QWERTY key which has "/" and
"?" ends up having "." and "," in Russian ЙЦУКЕН.
So if we want to have that key still do vi
searches, we need to rebind "." and ",", but
that would break command repeat and find-character
reversed repeat.

This also affects the number row. For example, "$"
on QWERTY is on Shift+"4", but Shift+"4" on ЙЦУКЕН
is ";", so we can't preserve the Shift+"4" muscle
memory without sacrificing find-character repeat.

Just using a one-time lookup for now instead of...

* ..binding directly to the English keyboard key
  (a macro definition), because I don't like the
  Emacs macro error message if the key it expands
  to isn't bound (could maybe figure out one day
  how to make the error message better).

* ..binding to a wrapper function that dynamically
  resolves the key in the current bindings, because
  I want `describe-key[-briefly]` to show me the
  actual function it resolves to rather (should be
  solved in the longer-term with function property
  that lets a command be transparent to an enhanced
  `describe-key[-briefly]`, which would also be
  very nice for window-state).

Skipping ex-commands for now.