.emacs.el: Russian keybindings: alias recursively
If a keybinding that would be aliased from QWERTY
to ЙЦУКЕН is itself a keymap, recurse into it.

This doesn't cover all keybindings, but it covers
almost all of them in practice (in particular, it
won't recurse into keybindings only reachable by
keys which are symbols on both QWERTY and ЙЦУКЕН.
But there's only one case I can think of where
this is a problem: my `space-map`).