.emacs.el: denoted-try: refactor into function
This is a second way to solve the bad interaction
between `denoted-try` and forms which evaluate to
function definitions as the `function` parameter
of `denoted-try`: turn `denoted-try` into a
function, its `function` parameter is evaluated
by the lisp runtime as normal.

Less surprises, composes better (`denoted-try`
can be used in `apply-partially`, etc).