.emacs.el: denoted-try: evaluate `function` parameter
There's a bad interaction between `denoted-try`
and forms which evaluate to function definitions
(such as `lambda-let`, which expands to an `eval`
of a `lambda`): since `denoted-try` is a macro,
it doesn't automatically expand and evaluate the
inner form.

Hitting the `function` argument to `denoted-try`
with an `eval` seems to work to fix this.