.emacs.el: allow `evil-repeat` to kill buffers
`evil-repeat` is coded with an opinionated safeguard
which errors out if the repeated operation tries to
kill a buffer. Personally I would think that when a
person tries to repeat a buffer-killing command,
they probably intend to kill a buffer, and that any
problematic buffer kills should be guarded in a more
general way (such as the "buffer {{name}} modified;
kill anyway?" prompt), but maybe that's just me.