.emacs.el: replace `unpack` and `uncons` with `pcase`
It's time. My fluency in Lisp/Elisp has grown
enough that I can now casually read and think
the pcase patterns for unpacking. Benefits:

* Standard construct which is familiar to many,
  is more likely to get tooling support, etc.
* Less complexity/tangling/coupling within my
  stuff: easier to pull out for reuse.
* The clarity of using `pcase-let*` vs `pcase-let`
  for explicitly calling out data dependencies.
* A couple of my `let-unpack` lines were unpacking
  an empty list just to avoid another `let` and
  indent level; `pcase-let` can do this cleaner
  since it's a superset of `let`.
* `_`-prefixed names aren't bound at all! Nice.

The downsides still apply: pcase is a lot less
readable in these cases, because

* the names `unpack` and `uncons` are more
  narrowly self-descriptive, and
* for these simple destructurings, pcase has
  far worse signal-to-noise ratio.

Honestly, I might even revert this commit, but
I'll give it a try for a while and get a sense
of how much mental overhead I spend on reading
pcase assignments.

(Aside: obviously the actually correct solution
is at the language level, to make all built-in
assignment forms support pattern destructuring.
Just let people write pcase patterns or similar
in `let`, in function parameters, in `setq`,
and so on.)