.emacs.el: fix git keybinds for unstaged files
When a file is inside a repo but not tracked
by that repo, `vc-deduce-backend` reports no
backend.

I suppose the proper move, perhaps, is to use
`vc-deduce-backend` on the file's directory,
in that case, but

1. I don't call `vc-deduce-backend` directly,
   I call `vc-repo-root`, and
2. I don't have enough reason to think that
   advising `vc-deduce-backend` to fall back
   from the file to its directory would be a
   good idea in the general case.

So I might as well just use my `git-repo-root`
function directly, since these commands are
already hardcoded to use Git.

(This fix is not applied to " vb", because
`vc-annotate` (`git blame` only makes sense
for staged files, and the current behavior
is better for falling into my custom better
error behavior where a separate buffer is
still popped.))