.emacs.el: make iterators introspectable
This fills a gap in generator.el functionality
by adding an `iter-p` which checks if an object
is a generator iterator.

This is done by advising `cps-generate-evaluator`,
which produces the generator body form which
evaluates to the iterator that the generator
function returns. The advice wraps the returned
form in a `(cons 'iterator ...)` form, ensuring
that the actual iterator is wrapped in a cons
cell with an `iterator` marker symbol.

The reason for having `iter-{next,close}` tolerate
iterators that are not wrapped with a marker is
that there is an initialization race: by the time
this code runs, other code might have already
define generators using the unadvised code.

(The only code this could break is code relying
on the representation of generator iterators,
which are currently just raw closures, and
Emacs/Elisp documentation says not to rely on
the representation of those.)