.emacs.el: make Enter not insert newline in non-inserting vi states
This fixes a nuisance where Enter inserts a newline
in modes that aren't supposed to insert text, such
as vi normal state in a plaintext file's buffer.

Since the intent is that Enter does nothing rather
than inserting a newline, the new `fixed-newline`
command also removes itself from Evil's repeat
information in those cases. (Debatable/unclear if
it would be better to just mark the command as not
"."-repeatable.)

(Out-of-the-box, Enter in vi/Evil is a motion, but
I have Enter in vi/Evil be transparent to the
underlying Enter binding, which is great for other
buffer types but not for "fundamental-mode" and
the like.)