.emacs.el: y-or-n-p: localize Russian mapping
Instead of mapping keyboard buttons by position
(which turned out to be the opposite of what I
intuitively expected once I forgot I did it), use
initials of the words "да" (yes) and "нет" (no).

This *does* run the risk of muscle memory for a
key sequence ending in a "yes" answer to hit "н"
since they're on the same key. But it seems that
I don't mentally move through "yes or no" prompts
the same way that I move through vi-style keybinds.
I seem to actually switch into language-coupled
thought for those - and if I'm intentionally in a
Russian keymap, that will probably be Russian.