.emacs.el: git-add-new: allow adding ignored files
I probably always want this behavior when I am
adding an individual file - if I'm saying to add
a file, especially when the keybind is four keys
of which one is a modifier, it's a pretty safe
bet that yes, I actually mean it.

But

1. for directories I definitely don't want it,
   since that could catch arbitrary junk;

2. I probably want to *know* if a file I am adding
   is caught by an ignore rule - either to adjust
   the rule, or to reconsider.

So I'm putting the "--force" flag behind the prefix
argument.