y-or-n: print default "n" as-if input
Thanks to the utter awfulness of

1. terminal design (no escape sequence to
   move cursor to end of printed line, no
   portable way to save and restore cursor
   position, etc),
2. shell redirections (can't capture stdout
   and stderr into separate variables, so
   it's hard to implement a clean+robust
   read-single-character shell command),
   and
3. other reasons blocking other possible
   solutions,

there's no good 100% correct way, let alone
a reasonably portable (even just modern-day,
practical portability!) way to do this. But
going up by one line and re-printing the
entire line (full prompt, the "n", and a
line feed) is 100% portable, and is 100%
correct-enough (only misbehavior is printing
a duplicate line, if the prompt was long
enough to line-wrap).