rmg: detect untracked files
I'm choosing to not detect untracked empty
directories for now, because probably those
are faintest whisps of ideas not worth
detecting before wiping out a repo's folder
(and this is consistent with `git status`).

If I ever have a system which stores a lot
of information as empty directories, and
has realistic odds of ending up in my repo
usage, I'll add `--directory` to the
`git ls-files` arguments.