.emacs.el: sorted-hash-table
Crude, simple version, based on a doubly-linked
list, so inserts are O(n)! Luckily, my use-case
so far is shell/REPL history, where n < 10k and
inserts are predominantly on the very tips (new
history entries are added at the start, sync'ed
history entries are usually at most just a few
entries from the start, and old history entries
during initial load are added at the end), so I
expect this will be good enough for a while.

(Note: the implementation uses my `bihash` as a
convenience for managing references from `dlist`
nodes back to their keys, but sorted-hash-table
is a uni-directional mapping, not a bijection;
the values never collide in this case because a
new dlist mode wraps every inserted value and
the cons cells are never `eq`.)