.emacs.el: add wait-for-file helper
One step up (and a small stutter back) from
a simple crude sleep: sleep in little bursts
in a loop until the file appears. Terrible,
but good enough for day-to-day use, and both
the UX and reliability is better than the
single sleep. (The ideal would be to set up
a file system watch, and check if the file
exists after setting up the watch just in
case it got created while executing the code
setting up the watch. But that's too much
time/effort for now.)
least the UX is better than a sleep that's
long enough to be reliable.