.emacs.el: window-state: try swapping "F" and "b"
I don't find myself using bury in other windows
as often as I find myself using unbury in the
current window... So I'm going to try this for
a while and see how I like it. (So now "sf" and
"sb" cycle the buffer of the current window,
while "sF" and "sB" cycle the buffer of the
window targetted by a motion.)

If I'm still unsatisfied with this, I might just
go back to the simplicity of matching vi more:
"f{{motion}}" for bury, "b{{motion}}" for unbury,
so "ff" or "F" and "bb" or "B" would also work
to bury/unbury current. This was how I had it in
the first place, iirc. The extra keystroke felt
annoying in the thick of testing and early use,
but maybe the consistency/memorability gains are
worth it.