.emacs.el: window-state-kill+delete operator
This is a combination I've been wanting for
a really long time, but for a mix of reasons
(need not quite strong enough, not seeing a
satisfying keybinding for it, and stuck on
this idea of implementing this as operator
combinations like "sdr{{motion}}").

For now, I'm just deleting the window instead
of reusing the delete operator's function,
since there's little sense in copying the buffer
that ends up in the window after the kill, and
and I don't yet have anything implemented to
restore a killed buffer so it's pointless to
copy the buffer just before killing it.