.emacs.el: window-state: "x" for kill+delete
(I'm going to move swap to "r" in an upcoming
commit, rationate in that commit message.)

Kill buffer + delete window is conceptually
similar to "x" in vi: it deletes one thing,
and it does not replace it with anything.

It's also mnemonic with "x" as the universal UI
icon for closing windows (closing a window in a
typical window manager is analogous to an Emacs
buffer kill with Emacs window delete: both the
window and the thing that was running/displaying
in that window are gone).