.emacs.el: window-state: "r" for swap
Swapping two windows is conceptually similar
to "r" in vi: it replaces the thing under
cursor ("point" in Emacs terms) with another
thing (specified immediately after). Not the
most precisely-fitting anology, but swapping
is more analogous with "r" and with "x".

I was originally using "x" for swap because
it was mnemonic with "exchange" and the
visual of "x" kinda looks reminiscent of
swapping. But that's not strong enough to
win now that there's a great window state
binding for "x" which is more analogous to
both vi an window manager UI conventions.

And since window state now has another
binding for killing buffers in a target
window (incidentally, the new "x" binding),
"r" is free. Plus, the mnemonic/rationale
analogizing buffer kill to vi's replace was
always a stretch.