.emacs.el: window-state-kill+delete; better error behavior
If the window is a sole ordinary window, the
error message from `delete-window` is utterly
pointless. Okay, great, it's the only window,
the user probably knows and just wanted to
kill the buffer. When there's only one window,
your OS window manager doesn't announce an
error from closing that window. Same concept.
When there's only one window in the selected
frame, the normal window managet analog to
"sxx" is closing a full-screen window.

(In a proper world, `delete-window` would
raise programmatically distinct error types
rather than `error` where the only distinct
feature is the prose message meant for humans.
But since that's missing, I have to do the
check ahead-of-time to decide whether to call
`(evil-)delete-window` instead of just catching
the error. And since I still want the native
error if there's an attempt to delete the
minibuffer, I also let the call go through
if `minibufferp` returns true.)