.emacs.el: window-state: rename target-window-prefix to target-prefix
`target-window-prefix` is the correct name
outside the context of window-state, matching
`same-window-prefix` and `other-window-prefix`.

But in the context of everything being a window,
and when there's no proper namespacing besides
prefixing the name with "window-state-", the
repetition of "window" in the name is a little
redundant and maybe a net signal-to-noise loss.
Unabbreviated self-describing/self-explaining
code names are good, but verbosity isn't free.