.emacs.el: new mpv-ipc code
Refactoring to reuse a persistent socket
directly from Emacs instead of executing
a socat each time.

This code handles opening the socket (in
Emacs, represented as a "process") and the
encoding/buffering/decoding IPC details.

This does does not manage the mpv process,
and instead is written to make it as easy
as possible to plug it into any approach:

* We can just use the socket path directly.
* If writing code that is starting the mpv
  process, we can extract the socket path
  from mpv's argument list (possibly an
  existing process object's argument list).

The second point elegantly allows the caller
to simultaneously

1. fully control program name, argument list,
   and how the process is started and managed
   (for example, it can even be launched
   inside an Eshell buffer), and yet
2. not do anything at all extra to pass the
   socket path into this code - the mere mpv
   argument `--input-ipc-server=...`, which
   might not even be in your code, can be the
   single source of truth.

This code also avoids having just one global
instance of anything: any number of `mpv` IPC
sockets can be worked with concurrently.