.emacs.el: wait-{while,until}: refactor + fix
There was a mistake which only showed up when
`check-interval` was a function that eventually
returned nil, which was hidden by the behavior
of lists in the case of finite lists: the right
behavior was to check one (last) time when the
next wait is `nil`. But before this commit, a
`check-interval` function would cause one extra
check: it would eval `test`, get a nil wait,
and eval `test` again.

Also, here's a better+fixed description:

    (wait-while test check-interval)

will

1. evaluate `check-interval` once, before
   the first iteration, then it will
2. evaluate `test` at the start of each
   iteration - if it is false, return
   `nil`, otherwise:
3. get `seconds` for `sleep-for`, based
   on what `check-interval` evaluated to:

   * if `check-interval` evaluated to a
     function, call that function with no
     arguments and use its return value
   * if `check-interval` evaluated to a
     list use the next element from it (or
     nil if there are no more elements)
     (`wait-while` neither mutates nor
     copies the list's cons cells);
   * if `check-interval` evaluated to
     anything else, use it as-is;

4. if `seconds` would be `nil`, return the
   result of the last evaluation of `test`.