.emacs.el: don't enter visual state on link touch
In Termux, I often find it convenient to click
on links in Help buffers and such. Somehow that
results in the originating buffer being put into
visual state. Since Termux has its own selection
feature on press-and-hold (and a scroll feature
that takes precedence for dragging after a press
without a hold), Evil's/Emac's text selection
on mouse drag is redundant on Termux anyway.