.emacs.el: music-map: move "d" to "x"
It seems that "x" is more intuitive for me
here. Maybe because it also has a stronger
UI association with the same action in media
players. (Or maybe because I tried "x" first,
used it a few times, and then went to sleep
immediately after swapping it to "d", so my
brain had more opportunity to lock in "x".)