.emacs.el: histdir-repl: micro-optimization
This fixes incorrect use of keywords supported
by `evil-define-operator`: the keywords must
go before the body, and the `(interactive ...)`
form is part of the body. Where I had them,
they were ignored by Evil and were just
expressions in the body with no side-effects
whose value was unused.

This wasn't causing any problems, so it would
have been valid to just delete these keywords,
but once fixed, they technically optimize out
a redundant movement of the point, and if I
wrote it correctly to begin with I wouldn't
delete them now, so might as well just fix it.