.emacs.el: music: vi-style delete operator on "d"
The big powerful thing here is that native
familiar vi motions are used to move over the
lines whose songs are supposed to be deleted
from the playlist. Instead of inventing new
vi-like motions, this is *just* an operator,
and it Just Works with all vi-style muscle
memory, mental skills and habits, etc.

Your brain-to-muscle interface doesn't change
when you pick up chains instead of pencils.

There are benefits to having the same motions
for everything, and just mapping standard
text editor motions to the underlying objects
by what the motions "touched" / covered.