.emacs.el: music: try "H"/"L" for previous/next
Analogous to how I've been using H and L for
previous/next for things like page history
(such as in Info and Help buffers).

The normal text-editing binding of "H" (the
"J" of stock vi) doesn't apply when playing
media: can't hoist up previous line to join
the current one. (..unless I bind that to
"stop looping this song, let it finish, and
start looping next song"... that's somewhat
analagous, and is the most analogous action
to the text editing binding of this key.)