.emacs.el: refactor music-{next,previous}
Follow the idiom of "next" implementing both
directions of the operation, and "previous"
just calling "next" with negated count.

This is of neglible/debatable benefit for a
function this small and simple (hence why I
didn't do it at first), but it's a good
pattern/habit/practice - compare the volume
"+"/"-" bindings I just added: there, the
trade-off is much more obvious. So I might
as well be consistent.