.emacs.el: mpv-ipc: add mpv-ipc-set helper
Mostly just for symmetry with `mpv-ipc-cycle`
and `mpv-ipc-add`, and setting forms such as
`setq`.

Another compelling reason is that mpv's IPC
`set` seemingly always requires numbers to be
passed in as strings (unlike `add`), which is
an API quirk no one should ever have to
remember or write boilerplate for.