.emacs.el: enable making commands not disturb vertical movement
Astoundingly, the core logic in `simple.el` is
hard-coded to only preserve the temporary goal
column if the last commands were exactly one
of `{next,previous}-line`.

This commit injects what's obviously objectively
the correct behavior: making it possible to make
any command "transparent to motions" (so that if
they are called between motions, it's still as
if we are in the same motion).

(The ideal behavior might be even more granular;
for example, a command might be sufficiently
different depending on prefix argument or thing
under point that it would want to specify if it
should be transparent to motion on a call by
call basis, but in principle that's now doable
by having command `foo` do something like
`(setq this-command 'foo-qux)` in some of its
branches, and only adding `foo-qux` into
`temporary-goal-column-preserving-commands`.)