.gitconfig: reduce detached head verbosity
I know what a detached head is. Usually if
it happens, I'm doing it on purpose, and I
want shorter text to *concisely* see which
parts of terminal scrollback were in
detached head state.

There's a small chance I'll miss it, but
I think I'll usually notice the concise
message without the screenful of warning.