.emacs.el: fix Enter in CRM right after separator
My improved CRM behavior was supposed to treat a
trailing separator (when hitting Enter) to mean
that we don't want to select whatever the blank
input after that last separator would select.

But that implementation had a bug which caused
the last choice (before the trailing separator)
to get expanded to a match, instead of being
used as-is (it would already have been expanded
to a match if desired at the time that the
separator was inserted, by hitting Comma or Tab,
so this was just wrongly expanding choices that
were intentionally not matches, with expansion
explicitly bypassed by hitting Alt+Comma).