.emacs.el: binary-tree-array
A set of functions for working with an array
representation of ("complete") binary trees.

In particular, it takes care of the hard part
of mapping the conceptual index (in the order
that the tree abstracts represents - the
indexes matching an inorder traversal) to the
internal concrete index of the array.

(I am sure this has already been done, but I
was unable to find it now and don't remember
ever seeing it.)

I want this this in-place for implementing
higher-performance sorted data structures.

Best I have on-hand right now is what I use
for history - a sorted doubly-linked list and
two hashmaps for bi-directional lookups from
keys to list links and back. This gives me O(1)
inserts on the edges and O(1) deletions, but
O(n) inserts in the middle, and O(n) search.

That's fine for the history use-case, where
the O(n) case is only ever hit if old entries
are newly added at runtime (if a device syncs
history after being offline for some time, but
even in those cases n is normally tiny and much
smaller than the full history - out-of-sync
history is rarely more than a few hours old,
and only needs to skip over history entries
which are newer than that).

But if I ever need something else, this will
help with implementing other techniques - in
particular, I am eyeing/considering doing what
the Python package `sortedcollections` does,
and the Jenks index is an array representation
of a complete/perfect binary tree.