bin/l: removed
On further thought, I don't really use `l`
either, because in Emacs on Termux, `eza`
output causes drag-scrolling to lag (and
I don't find myself wanting shorthand for
`ls -l`: its output is less valuable, and
in the less common cases where I do want
it, it's short/easy/quick enough to type).

Might as well remove it. One day, I might
add it back, or I might wrap/upgrade the
`ls` that's built into Emacs to do some
of the best parts of Exa/Eza (the biggest,
for me, it that the distinct colors for
each column really ease visual parsing,
especially when lines wrap, as is the
norm on narrow phone screens).