.emacs.el: mpv-ipc: tolerate event messages
Normally, there shouldn't be event messages,
since the first thing we do is disable all
events - but there is a race condition: an
event message can happen before disabling.
This commit plugs that hole, by making the
code that consumes an IPC message discard
and repeat whenever it consumes an event
rather than a reply.