.gwsl: fix sloppy quoting omission
For approximately the first time in almost
a decade, I have slipped up on shell script
quoting: the `export` didn't cause me to
flip from special (no quoting of expansions
needed) assignment rules to regular rules.

Funny to see the best practice and all that
I preach about cognetic openings vindicated
so clearly - even with my decade or so of
habitual review of the relevant cognition,
I slip up, and in fact the slip-up is
basically an inevitable result of my mental
growth/healing (I have grown/healed to give
higher priority to me getting my
bigger-picture goals done, to velocity
and my human-level efficiency).

I think this is the moment that I finally
let go of this aspect of my obsession with
/bin/sh portability: no more backticks for
the shells that lack `$(...)`, no more
omitting quotes for the special spots of
shell grammar that don't do splitting or
globbing (a habit originally justified by
consistency with it being impossible to
quote backticks which contain quotes).