.emacs.el: eshell/r: suppress PAGER
1. Emacs+vi is mostly the ideal pager anyway.
2. The only pager use-case I have from a
   wrapped REPL is Python's `help`, but
   I often find myself wishing `help` didn't
   use a pager, because I want to consult
   the help text more than once, or while
   I am editing a REPL input.
3. histdir-repl's vi-style "insert"/etc have
   some issues when the beginning-of-input
   and end-of-input inputs cause a terminal
   redraw, which doesn't happen in a normal
   REPR but *does* happen if `less` is used
   (`less` has no keybind for Ctrl+A but does
   have a default Ctrl+E keybind that scrolls
   down by one line).

For my personal purposes, it might be righter
to just solve this with Python's `usercustomize`
until I have any other REPL trying to use `PAGER`.
In the long term, there probably needs to be a
solution deeper in histdir-repl, but I'm still
contemplating what that should look like (and
even whether that should suppress paging or just
somehow handle the terminal redraw). So this
commit implements a middleground, suppressing
paging at the level of my Eshell `r` entry point
to histdir-repl.