.emacs.el: make markdown-mode fontify "directives"
Fenced directives are a growingly popular
overload of fenced code block syntax, which
allows extending Markdown with arbitrary
custom blocks, in a way that's extremely
easy for Markdown parsers to support.

I am increasingly using them in my personal
infrastructure, and since they're a great
general-purpose solution to Markdown's
extensibility problem, I expect that trend
to continue.

This patch makes directives' blocks not
fontify as code blocks (thus keeping them
free for consequent fontification of other
Markdown syntax, such as bold, inline code,
and so on).

A directive is just a code fence with a name
wrapped in curly braces:

    ```{foo}
    Arbitrary contents. Directive decides
    how its contents are parsed.
    ```

Some Markdown parsers support `:::` fences
instead of backtick fences, in large part
to avoid the need for syntax highlighting
logic changes like this patch. So in the
future I might roll this back and just use
`:::` fences for directives that contain
nested Markdown, but for now I want to see
how far I can get with just one fence type.