.emacs.el: dlist: add "reverse" cdr/nth operations
Been holding off on adding proper abstractions
for headward link traversal/setting, since I
couldn't think of a good way to name them.

From scratch, I'd use the idiomatic word "tail"
and the opposite of tail is obviously *body*
(not "head" - the head of a linked list is
a single node, we want the word for all of
the links headwards of the current node: all
all the nodes between tail and head).

Trying to stick to other conventions, I'd
maybe use "next" and "previous" or "left"
and "right".

But I wanted to stick with something based
on Lisp's native "cdr" (the cheeky "ccr" is
the obvious name for an alternate "cdr",
leaving "cbr" as an alternate "car", but
that's too cutesy). I imagine eventually I just
give up on trying to fit into Lisp's horrible
ancient-implementation-detail naming, but for
now this whole reverse-flag approach seems
sorta reasonable.