.emacs.el: make vi `%` work at end of buffer
Longstanding Evil nuisance: when a lisp data
structure is logged to the Messages buffer,
I usually go to the Messages buffer (point is
at end of buffer) and hit `%` to get to the
beginning of the structure, but get an error
instead. This commit acommodate's Evil's buggy
assumption by first moving the point off the
end by one character.