.emacs.el: dlist-cpr: previous/headward methods
This commit adds functions for the "previous"
link of `dlist` nodes, analogous to the "cdr"
functions for the "next" link. This abstracts
the implementation detail of using `caar` and
`(setcar (car ...) ...)` with proper APIs,
finally.

The "cpr" Naming is only half-serious/permanent.

Just as there was once `car` and `cdr` at the
hardware level, there once was a `cpr`. The
IBM 704 had these instructions for extracting
the Contents of the {Address,Decrement,Prefix}
part of a "Register" (memory address, modern
terminology had not yet settled).

This leads to a brilliant/awful cheeky naming
idea: "prefix" semantically fits getting the
"previous" link of a doubly-linked list, since
the nodes in front of the current node are by
definition a prefix. (And the "P" is mnemonic
for "previous", but that's a weaker reason.)

(Part of why I love this so much is that there
is also a `ctr`, for the "Tag" part, which is
yet another word/letter that could be reused in
semantically/mnemonically fitting ways: such as
for additional metadata.)

P.S. I had this idea months ago, but put off
implementing it until I had enough need/want
for "headward/previous link" functions/names.

P.P.S. I will probably implement proper naming
at a later time (a simple `next` and `previous`).
I might entirely drop the `car` and `cdr` names
at that point too. So for example, instead of
`dlist-nthcdr` there will be `dlist-nth-next`,
or `dlist-next` (renaming `dlist-cdr`) will gain
an `&optional n` parameter. But for now this is
a fun little playful experiment that sparks joy,
and I predict low odds of myself tripping over
it or anyone having to deal with this naming.