.emacs.el: dlist-nth: make negative n go headwards
For native Lisp lists, `nth` and `nthcdr`
behave as if `n` is 0 when `n` is negative.
And for singly-linked lists that's elegant,
convenient, and sensible (in as much as all
forms of "define errors out of existence"
are - it is still a tradeoff because it can
turn mistakes into silent misbehavior).

But for a doubly-linked list, there is another
valid and useful use of negative `n`: going
backward (headwards - towards the head link).
For `dlist-nthcdr` it's very debatable: it is
reasonable to think of "cdr" as always "next",
always the tailward/forward link. But there's
no such counter-argument for `dlist-nth` (it
only makes sense for singly-linked lists as a
result/enablement of elegantly implementing
`(nth ...)` as merely `(car (nthcdr ...))`).

Now that there's proper API support for "cpr"
as well as "cdr", cleanly splitting the two
sides (previous and next) of node/link-oriented
accesses, confrontingly leading to a situation
where it's totally arbitrary to make `dlist-nth`
a wrapper for `dlist-nthcdr` when `dlist-nthcpr`
exists, and it makes sense to let `dlist-nth`
take on the more natural design/semantic of
being able to go in either direction.